extract bundled frontend adapter
