--- src/config/constants.ts
@@ -1,60 +1,0 @@
-// generated by src/config/constants.gen.ts
-
-// note: in environments where bundle size is important,
-// prefer this module to using /src/fs/mime.ts when possible
-
-// this list can be mutated after importing
-export const DEFAULT_ASSET_PATHS: string[] = [
-	'webp',
-	'tif',
-	'tiff',
-	'svg',
-	'png',
-	'jpg',
-	'jpeg',
-	'ico',
-	'gif',
-	'bmp',
-	'apng',
-	'ttc',
-	'woff2',
-	'woff',
-	'ttf',
-	'otf',
-	'mp4',
-	'mp4v',
-	'mpg4',
-	'mpeg',
-	'ogv',
-	'webm',
-	'm4a',
-	'mp4a',
-	'flac',
-	'mid',
-	'midi',
-	'mp3',
-	'oga',
-	'ogg',
-	'weba',
-	'wav',
-	'csv',
-	'js',
-	'mjs',
-	'html',
-	'htm',
-	'css',
-	'txt',
-	'log',
-	'epub',
-	'zip',
-	'gz',
-	'ogx',
-	'pdf',
-	'xhtml',
-	'xml',
-	'jsonld',
-	'json',
-	'map',
-];
-
-// generated by src/config/constants.gen.ts
