+++ src/gen.task.ts
@@ -3,7 +3,7 @@ import {TaskError} from './task/task.js';
 import {red, green, gray} from './utils/terminal.js';
 import {runGen} from './gen/runGen.js';
 import {loadGenModule, checkGenModules, findGenModules} from './gen/genModule.js';
-import {printPath, printMs, printError, printTiming} from './utils/print.js';
+import {printPath, printMs, printError, printTimings} from './utils/print.js';
 import {resolveRawInputPaths} from './fs/inputPath.js';
 import {plural} from './utils/string.js';
 import {createStopwatch, Timings} from './utils/time.js';
@@ -116,9 +116,7 @@ export const task: Task<TaskArgs> = {
 				} input file${plural(genResults.successes.length)}`,
 			),
 		);
-		for (const [key, timing] of timings.getAll()) {
-			log.trace(printTiming(key, timing));
-		}
+		printTimings(timings, log);
 		log.info(`🕒 ${printMs(totalTiming())}`);
 
 		if (failCount) {
